refactor : move search grid to another file
