+++ client/src/components/recipes/recipes.tsx
@@ -1,9 +1,10 @@
 import { useState, useEffect } from 'react';
-import { FormControl, createTheme, ThemeProvider, List, ListItem, ListItemText, Divider, Drawer, Box, CssBaseline, Select, TextField, AppBar, Toolbar, Typography, IconButton, Menu, MenuItem, Checkbox, FormControlLabel, Grid, Card, CardActionArea, CardMedia, CardContent } from '@mui/material';
-import { SearchResult, RecipeSearchResults, RecipeSearchParams } from '../../types/search_types';
-import { useNavigate } from "react-router-dom"
+import { FormControl, List, ListItem, ListItemText } from '@mui/material';
+import { SearchResult } from '../../types/search_types';
 import { SelectInput } from './select';
-import {SearchResultGrid} from'./searc-result-grid'
+import { SearchResultGrid } from './search-result-grid';
+import SideDrawer from '../general/side_drawer';
+//import MultipleSelectWithCheckboxes from '../general/multi_select_checks'
 
 const BASE_URL = "http://localhost:3000/api/v1/search";
 
@@ -18,6 +19,11 @@ const Recipes: React.FC = () => {
     const [intolerances, setIntolerances] = useState("None");
     const [vegetarianOnly, setVegetarianOnly] = useState(false);
     const [searchText, setSearchText] = useState("");
+    const [searchCriteria, setSearchCriteria] = useState("");
+
+    const toggleSidePanelDrawer = (open: boolean) => {
+        setSidePanelOpen(open);
+    };
 
     const cuisineList: string[] = ['All', 'African', 'American', 'British', 'Cajun', 'Caribbean', 'Chinese',
         'Eastern European', 'European', 'French', 'German', 'Greek', 'Indian', 'Irish', 'Italian',
@@ -34,97 +40,61 @@ const Recipes: React.FC = () => {
 
         return (
             <>
-                <Drawer anchor="right" open={sidePanelOpen} onClose={() => setSidePanelOpen(false)}>
-                    <FormControl variant='outlined' size="small" style={{ width: '100%' }}>
-                        <List sx={{ width: 250, mt: 2, mb: 2 }}>
-                            <ListItem>
-                                <ListItemText primary="Search Filters" />
-                            </ListItem>
-                            {/* <SelectCuisine cuisine={cuisine} onChange={(newValue) => setCuisine(newValue)} /> */}
-                            <SelectInput
-                                id='cuisine'
-                                name='cuisine'
-                                value={cuisine}
-                                label='Cuisine'
-                                onChangeHandler={(newValue) => setCuisine(newValue)}
-                                options={cuisineList}
-                            />
+                <FormControl variant='outlined' size="small" style={{ width: '100%' }}>
+                    <List sx={{ width: 250, mt: 2, mb: 2 }}>
+                        <ListItem>
+                            <ListItemText primary="Search Filters" />
+                        </ListItem>
+                        <SelectInput
+                            id='cuisine'
+                            name='cuisine'
+                            value={cuisine}
+                            label='Cuisine'
+                            onChangeHandler={(newValue) => setCuisine(newValue)}
+                            options={cuisineList}
+                        />
 
-                            <SelectInput
-                                id='mealtype'
-                                name='mealtype'
-                                value={mealType}
-                                label='Type'
-                                onChangeHandler={(newValue) => setMealType(newValue)}
-                                options={mealTypeList}
-                            />
-                            <SelectInput
-                                id='intolerances'
-                                name='intolerances'
-                                value={intolerances}
-                                label='Intolerances'
-                                onChangeHandler={(newValue) => setIntolerances(newValue)}
-                                options={intolerancesList}
-                            />
-                            {/* <SearchText searchText={searchText} onChange={(newValue) => setSearchText(newValue)} /> */}
-
-                            {/* <ListItem>
-                                <FormControlLabel
-                                    control={
-                                        <Checkbox
-                                            checked={vegetarianOnly}
-                                            onChange={handleVegetarianOnlyChange}
-                                            color="primary"
-                                        />
-                                    }
-                                    label="Vegetarian Only"
-                                />
-                            </ListItem> */}
-                        </List>
-                    </FormControl>
-                </Drawer>
+                        <SelectInput
+                            id='mealtype'
+                            name='mealtype'
+                            value={mealType}
+                            label='Type'
+                            onChangeHandler={(newValue) => setMealType(newValue)}
+                            options={mealTypeList}
+                        />
+                        <SelectInput
+                            id='intolerances'
+                            name='intolerances'
+                            value={intolerances}
+                            label='Intolerances'
+                            onChangeHandler={(newValue) => setIntolerances(newValue)}
+                            options={intolerancesList}
+                        />
+                        <ListItem>
+                            <button onClick={() => setSidePanelOpen(false)}>Search</button>
+                        </ListItem>
+                    </List>
+                </FormControl>
             </>
         )
     };
 
-    
+
 
     const searchParameters = (): string => {
         const searchParams = new URLSearchParams();
-        //console.log("******* params", params)
 
         if (cuisine !== "All") searchParams.set('cuisine', cuisine);
-        //   if (params.excludeCuisine !== undefined) searchParams.set('excludeCuisine', params.excludeCuisine);
-        //   if (params.diet !== undefined) searchParams.set('diet', params.diet);
         if (intolerances !== "None") searchParams.set('intolerances', intolerances);
-        //   if (params.includeIngredients !== undefined) searchParams.set('includeIngredients', params.includeIngredients);
-        //   if (params.excludeIngredients !== undefined) searchParams.set('excludeIngredients', params.excludeIngredients);
         if (mealType !== "All") searchParams.set('type', mealType);
         if (searchText !== "") searchParams.set('titleMatch', searchText);
-        //   if (params.maxReadyTime !== undefined) searchParams.set('maxReadyTime', params.maxReadyTime.toString());
-        //   if (params.sort !== undefined) searchParams.set('sort', params.sort);
-        //   if (params.sortDirection !== undefined) searchParams.set('sortDirection', params.sortDirection);
-        //   if (params.minCarbs !== undefined) searchParams.set('minCarbs', params.minCarbs.toString());
-        //   if (params.maxCarbs !== undefined) searchParams.set('maxCarbs', params.maxCarbs.toString());
-        //   if (params.minProtein !== undefined) searchParams.set('minProtein', params.minProtein.toString());
-        //   if (params.maxProtein !== undefined) searchParams.set('maxProtein', params.maxProtein.toString());
-        //   if (params.minCalories !== undefined) searchParams.set('minCalories', params.minCalories.toString());
-        //   if (params.maxCalories !== undefined) searchParams.set('maxCalories', params.maxCalories.toString());
-        //   if (params.minFat !== undefined) searchParams.set('minFat', params.minFat.toString());
-        //   if (params.maxFat !== undefined) searchParams.set('maxFat', params.maxFat.toString());
-
-        console.log("**searchParams1111**", searchParams.toString())
         return searchParams.toString();
     }
-
-
-
 
     const fetchRecipe = async () => {
 
         const response = await fetch(`${BASE_URL}?${searchParameters()}`);
         const data = await response.json();
-        console.log("****** client serach result =>", data.results);
         setSearchResult(data.results);
     };
 
@@ -133,23 +103,13 @@ const Recipes: React.FC = () => {
         console.log(" fetch recipe")
     }, [sidePanelOpen]);
 
-    // const handleCuisineChange = (event) => {
-    //     setCuisine(event.target.value);
-    // };
-
-    // const handleVegetarianOnlyChange = (event) => {
-    //     setVegetarianOnly(event.target.checked);
-    // };
-
-    // const handleSearchTextChange = (event) => {
-    //     setSearchText(event.target.value);
-    // };
-
     return (
-        <>  
-        <button onClick={() => setSidePanelOpen(true)}>Search Filter</button>
-            <FilterPanel />
-            <SearchResultGrid recipeList = {searchResult} />
+        <>
+            <button onClick={() => setSidePanelOpen(true)}>Search Filter</button>
+            <SideDrawer isOpen={sidePanelOpen} toggleDrawer={toggleSidePanelDrawer}>
+                <FilterPanel />
+            </SideDrawer>
+            <SearchResultGrid recipeList={searchResult} />
         </>
     )
 };
