+++ client/src/components/recipes/search-result-grid.tsx
